Fixed SW and NE bug
It was just a bad assignment of the BlendSpace animation node. They were set to the wrong animation and it made it look like it wasn't working. Also changed the naming of all the animations using numpad definitions. Now we can update the animations to change the texture instead of making it each a separate node.